refactor(github): extract repository & gist backup flows into helper
