+++ internal/vcs/github/repos.go
@@ -8,6 +8,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"strings"
 	"time"
 
 	"github.com/google/go-github/v59/github"
@@ -112,5 +113,53 @@ func (g *GitHubVCS) pullRepository(repoDir string) error {
 	}
 
 	log.Printf("Successfully updated repository at %s\n", repoDir)
+	return nil
+}
+
+// backupRepositories orchestrates the backup of all repositories for a user.
+func (g *GitHubVCS) backupRepositories(ctx context.Context, username string) error {
+	repos, err := g.listRepositories(ctx, username)
+	if err != nil {
+		return fmt.Errorf("failed to list repositories: %w", err)
+	}
+
+	log.Printf("Found %d repositories for user %s\n", len(repos), username)
+
+	// Create repositories directory
+	reposDir := filepath.Join(g.config.OutputDir, "repos")
+	if err := os.MkdirAll(reposDir, 0755); err != nil {
+		return fmt.Errorf("failed to create repositories directory: %w", err)
+	}
+
+	// Backup repositories concurrently
+	errChan := make(chan error, len(repos))
+	semaphore := make(chan struct{}, g.config.Threads)
+
+	for _, repo := range repos {
+		semaphore <- struct{}{} // Acquire semaphore
+		go func(r *github.Repository) {
+			defer func() { <-semaphore }() // Release semaphore when done
+
+			if r.Name == nil {
+				errChan <- fmt.Errorf("repository has no name")
+				return
+			}
+
+			if err := g.cloneRepository(ctx, r, reposDir); err != nil {
+				errChan <- fmt.Errorf("failed to backup repository %s: %w", *r.Name, err)
+				return
+			}
+
+			errChan <- nil
+		}(repo)
+	}
+
+	// Wait for all goroutines to complete
+	for i := 0; i < len(repos); i++ {
+		if err := <-errChan; err != nil {
+			log.Printf("Error during backup: %v", err)
+		}
+	}
+
 	return nil
 }
